MYR-507/MYR-515 Address Claude Review: file cap, and the mirror half-pair gap
Review verdict was APPROVE with no critical findings. Two items:

WARNING (blocks merge per CLAUDE.md): vehicles_list_handler.go hit 302 lines,
one over the 300-line cap. Moved newVehicleSummary + lastFourOfVIN out — but
into a THIRD file rather than vehicles_list_types.go as suggested, because
that would have pushed the types file to 304 and moved the violation rather
than fixing it. The trio is now handler (233, request flow) / types (228,
shapes) / projection (89, the mapping between them), all well under.

SUGGESTION: the half-pair assertion covered only lat-set/lng-nil. Added the
mirror row. The guard is a two-clause ||, so one permutation would still pass
with a clause dropped — exactly the edit a later refactor might make.

Not taken: the plausible()/normalize() double-TrimSpace. The reviewer called it
negligible and not worth changing on its own, and threading a pre-normalized
string through would trade two trivial allocations for a sharper API.
