+++ internal/telemetry/vehicles_list_handler.go
@@ -217,75 +217,6 @@ func (h *VehiclesListHandler) buildResponse(rows []VehicleCatalogRow, role auth.
 	return vehiclesListResponse{Items: items}
 }
 
-// newVehicleSummary builds the per-row wire shape for a catalog row under a
-// given role. grant is the caller's share capability set and is read ONLY on
-// viewer rows — pass the zero ShareGrant for an owner, who holds no grant.
-//
-// Shared by the owner list, the viewer merge (vehicles_list_viewer.go), and the
-// redeem response, so all three emit byte-identical rows for the same vehicle.
-//
-// now is the instant the MYR-491 setup derivation is judged against, passed in
-// rather than read here for the same two reasons as on the snapshot: one
-// response is one instant, and the rules are testable without waiting.
-func newVehicleSummary(v *VehicleCatalogRow, role auth.Role, grant auth.ShareGrant, now time.Time) vehicleSummary {
-	summary := vehicleSummary{
-		VehicleID:      v.ID,
-		Name:           v.Name,
-		Model:          v.Model,
-		Year:           v.Year,
-		Color:          v.Color,
-		LicensePlate:   v.LicensePlate,
-		VinLast4:       lastFourOfVIN(v.VIN),
-		Status:         v.Status,
-		ChargeLevel:    v.ChargeLevel,
-		EstimatedRange: v.EstimatedRange,
-		LastUpdated:    v.LastUpdated.UTC().Format(time.RFC3339),
-		Role:           string(role),
-		HasActiveRide:  v.HasActiveRide,
-		// MYR-316: resolved here so the precedence and the in-service gate
-		// are applied exactly once per surface (service_window.go).
-		ServiceEstimatedEndAt: serviceEstimatedEndAtWire(v.Status, v.ServiceETC, v.ServiceExpectedEndAt),
-		// MYR-342: emitted verbatim on BOTH roles. The mask, not this
-		// assignment, is what decides who sees it — and both allow-lists carry
-		// it, because a rider who cannot see that a shared car is paused finds
-		// out from a 409 instead.
-		RideShareEnabled: v.RideShareEnabled,
-		// MYR-507: emitted verbatim on BOTH roles, from the same column
-		// /snapshot reads. Identity, not telemetry and not privacy-bearing —
-		// see the mask allow-list for why a viewer sees it.
-		TrimLabel: v.TrimLabel,
-		// MYR-515: resolved here so the atomic-pair rule and the (0,0)
-		// sentinel collapse are applied exactly once per surface, in the same
-		// place the MYR-316 window resolves its precedence.
-		Location: newVehicleLocation(v.Latitude, v.Longitude),
-		// MYR-491: the SAME derivation the snapshot runs, over the same inputs.
-		// Calling it from both surfaces rather than copying the rules is what
-		// makes "the catalog and the detail sheet can never disagree" a
-		// structural property instead of a convention.
-		SetupState: deriveSetupState(now, v.Status, v.LastUpdated, v.SetupSchedule),
-	}
-	if role == auth.RoleViewer {
-		// DERIVED, not stored (MYR-369): the grant's flags decide the
-		// compatibility value a pre-MYR-369 client reads. A suspended grant
-		// never reaches here — the viewer's catalog query excludes it, so
-		// the vehicle is absent from the response entirely rather than
-		// present with some reduced value.
-		summary.SharePermission = grant.Permission().String()
-	}
-	return summary
-}
-
-// lastFourOfVIN returns the last 4 characters of a VIN. Empty input
-// yields an empty string; shorter-than-4 inputs are returned verbatim
-// (the contract guarantees 17-char VINs, but the helper is defensive
-// against test fixtures that pass shorter values).
-func lastFourOfVIN(vin string) string {
-	if len(vin) <= 4 {
-		return vin
-	}
-	return vin[len(vin)-4:]
-}
-
 // writeJSON marshals v as JSON with the given status code.
 func (h *VehiclesListHandler) writeJSON(w http.ResponseWriter, status int, v any) {
 	w.Header().Set("Content-Type", "application/json; charset=utf-8")
